+++ src/containerapp/azext_containerapp/_constants.py
@@ -6,6 +6,7 @@ MANAGED_ENVIRONMENT_TYPE = "managed"
 CONNECTED_ENVIRONMENT_TYPE = "connected"
 MANAGED_ENVIRONMENT_RESOURCE_TYPE = "managedEnvironments"
 CONNECTED_ENVIRONMENT_RESOURCE_TYPE = "connectedEnvironments"
+CUSTOM_LOCATION_RESOURCE_TYPE = "customLocations"
 
 MAXIMUM_SECRET_LENGTH = 20
 MAXIMUM_CONTAINER_APP_NAME_LENGTH = 32
@@ -20,6 +21,8 @@ CONTAINER_APPS_SDK_MODELS = "azext_containerapp._sdk_models"
 LOG_ANALYTICS_RP = "Microsoft.OperationalInsights"
 CONTAINER_APPS_RP = "Microsoft.App"
 SERVICE_LINKER_RP = "Microsoft.ServiceLinker"
+EXTENDED_LOCATION_RP = "Microsoft.ExtendedLocation"
+CONTAINER_APP_EXTENSION_TYPE = "microsoft.app.environment"
 
 MANAGED_CERTIFICATE_RT = "managedCertificates"
 PRIVATE_CERTIFICATE_RT = "certificates"