+++ src/serviceconnector-passwordless/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+0.3.9
+++++++
+* Support `--customized-keys` and make some improvements.
+
 0.3.8
 ++++++
 * Make some improvements.